% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CNV_catfishr_data}
\alias{CNV_catfishr_data}
\title{Example CNV data for CATFISHR}
\format{
An object of class \code{list}.
}
\source{
Generated for package example.
}
\usage{
CNV_catfishr_data
}
\description{
CNV input data used in the CATFISHR vignette.
}
\keyword{datasets}
