% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_mahal_output.R
\name{format_mahal_output}
\alias{format_mahal_output}
\title{Format RNA and CNV Mahalanobis distance outputs}
\usage{
format_mahal_output(mahal_RNA, mahal_CNV)
}
\arguments{
\item{mahal_RNA}{Output list from \code{\link[=calc_mahalanobis]{calc_mahalanobis()}} for RNA PCA space.}

\item{mahal_CNV}{Output list from \code{\link[=calc_mahalanobis]{calc_mahalanobis()}} for CNV PCA space.
outputs are kept. If \code{TRUE}, all cells from either output are kept and
missing values are filled with \code{NA}. Default \code{FALSE}.}
}
\value{
A data frame with one row per cell and columns:
\describe{
\item{sample_barcode}{Cell barcode.}
\item{Mahal_Dist_RNA}{Minimum squared Mahalanobis distance in RNA space.}
\item{Mahal_Dist_CNV}{Minimum squared Mahalanobis distance in CNV space.}
\item{RNA_ref_status}{If cell in RNA reference cluster.}
\item{CNV_ref_status}{If cell in CNV reference cluster.}
\item{RNA_ori_cluster}{Original RNA cluster assignments.}
\item{CNV_ori_cluster}{Original CNV cluster assignments.}
}
}
\description{
Combines the minimum Mahalanobis distances from RNA and CNV
\code{calc_mahalanobis()} outputs into a single data frame for downstream
mean shift clustering and plotting.
}
