% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RNA_catfishr_data}
\alias{RNA_catfishr_data}
\title{Example RNA data for CATFISHR}
\format{
An object of class \code{list}
}
\source{
Generated for package example.
}
\usage{
RNA_catfishr_data
}
\description{
RNA input data used in the CATFISHR vignette.
}
\keyword{datasets}
