% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mahalanobis.R
\name{select_pcs}
\alias{select_pcs}
\title{Select number of PCs using variance drop-off heuristic}
\usage{
select_pcs(pca_matrix, drop_threshold = 0.1, default_pcs = 30)
}
\arguments{
\item{pca_matrix}{A cells x PCs matrix of PCA embeddings.}

\item{drop_threshold}{Minimum percent variance drop between consecutive PCs.}

\item{default_pcs}{Number of PCs to use if no drop exceeds threshold.}
}
\value{
Integer number of PCs to use.
}
\description{
Select number of PCs using variance drop-off heuristic
}
